Develop fix

+++ src/main/java/com/screenerd/service/PostService.java
@@ -5,6 +5,8 @@ import com.screenerd.domain.Post;
 import com.screenerd.domain.User;
 import com.screenerd.repository.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 /**
@@ -21,17 +23,17 @@ public class PostService {
         if (post == null)
             throw new IllegalArgumentException("Post cannot be null");
         User author = post.getUser();
-        if (author != null) {
-            if (author.getId() == null)
-                userService.saveUser(author);
-            author.getPosts().add(post);
-        }
-        return postRepository.save(post);
+        postRepository.save(post);
+        author.addPost(post);
+        return post;
 
     }
 
     public void deletePost(Long id) {
         Post post = this.postRepository.findOne(id);
+        if(post == null){
+            throw new IllegalArgumentException("Can not delete inexisting post");
+        }
         post.getUser().getPosts().remove(post);
         this.postRepository.delete(id);
     }
@@ -44,8 +46,8 @@ public class PostService {
         this.postRepository = postRepository;
     }
 
-    public Iterable<Post> findAllPosts() {
-        return this.postRepository.findAll();
+    public Page<Post> findPage(Pageable pageable) {
+        return postRepository.findAll(pageable);
     }
 
     public Post findPostById(Long id) { return this.postRepository.findOne(id);}
